requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.3
django-filter==2.0.0
django-heroku==0.3.1
django-rest-swagger==2.1.2
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
openapi-codec==1.3.2
psycopg2==2.7.6.1
pytz==2018.7
requests==2.20.1
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
